Move timestamp utility functions into a separate api functions file
